changes applied on the orders controller

+++ final-project/Models/orderline.cs
@@ -5,7 +5,7 @@
         public int id { get; set; }
         public string itemname { get; set; }    
         public int itemquant { get; set; }      
-        public decimal itemprice { get; set; }  
+        public int itemprice { get; set; }  
         public int orderid { get; set; }
     }
 }
